app/requirements.txt
Django==3.2.18
django-filter==22.1
django-health-check==3.17.0
django_redis
django-permissions-policy==4.14.0
django-cors-headers==3.13.0
django-csp==3.6
django-db-schema-renderer==0.1.2
celery==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor
django-session-timeout-joinup==1.0.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
djangorestframework-xml==2.0.0
drf-extensions==0.7.1
drf-writable-nested==0.7.0
drf-spectacular==0.26.0
drf_extra_fields==3.4.1
uWSGI==2.0.22
psycopg2>=2.9.5
django-webpack-loader==1.7.0
requests==2.29.0
debugpy
Pillow==9.4.0
pillow-heif==0.13.0
sorl-thumbnail==12.8.0


redis==4.5.5
mozilla-django-oidc==3.0.0
PyJWT<2.4.0

device-detector==5.0.1
geopy==2.4.1
chardet==5.2.0
hashids==1.3.1
django-debug-toolbar==4.0.0
